fix: address PR review comments — bug fix, doc corrections, lint improvement
Bug fix
-------
config_storage: FNV sign-extension corrected in get_keyboard_id().  char
values 0x80–0xFF were sign-extended to 32 bits before the XOR step,
corrupting the keyboard-ID hash on signed-char targets.  Cast changed from
(uint32_t)(*str++) to (uint32_t)(uint8_t)(*str++).

Refactoring
-----------
keylayers: duplicated layer-count clamping logic extracted into a
get_effective_layer_count() static inline helper.  Both handle_valid_hash()
and keylayers_process_key() now call it, removing ~6 lines of repeated code.

at-ps2/keyboard_interface: GNU range-extension syntax
case INIT_READ_ID_1 ... INIT_SETUP: replaced with three explicit ISO C11
case labels, removing a non-portable GNU extension.

Documentation
-------------
usb_descriptors: @note Main loop only added to all four TinyUSB descriptor
callbacks (tud_descriptor_device_cb, tud_hid_descriptor_report_cb,
tud_descriptor_configuration_cb, tud_descriptor_string_cb).  Magic number
31 replaced with named constants DESC_STR_BUF_WORDS (32U) and
DESC_STR_MAX_CHARS (DESC_STR_BUF_WORDS - 1U) in three places.

at-ps2/common_interface: full Doxygen blocks added to atps2_send_command()
and atps2_setup_pio_engine() — all @param, @return, and @note entries,
including cleanup-on-failure behaviour and ISR restrictions.

at-ps2/mouse_interface: @note annotation corrected on all seven per-state
handlers from "Main loop only" to "IRQ-safe — called exclusively from
mouse_input_event_handler() which runs in ISR context."  @return void
added to each handler.

set1/scancode: stale inline comment updated to reflect the corrected
SC1_MAX_CODE value (was "211 decimal"; now "0xD8 / 216 decimal").

set123/scancode: stale break-range comment updated (0x81-0xD3 →
0x81-0xD8 (SC1_MAX_CODE)).  @param code, @param config, @return void, and
@note Main loop only added to all ten per-state handler Doxygen blocks;
handlers that suppress the unused config parameter document it as such.

set2/scancode: @param code and @return void added to all nine per-state
handler Doxygen blocks.

Tooling
-------
lint.sh: Check 19 regex broadened to match multi-word types such as
"unsigned int" and "unsigned long" (previously only single-token types
matched).  Spacing around = tightened to handle varied styles.
Explanatory comments added documenting the indentation heuristic, its
known edge case, and the LINT:ALLOW workaround.

tools/README.md: check count corrected (eighteen → nineteen).

+++ src/common/lib/config_storage.c
@@ -88,7 +88,7 @@ static uint32_t get_keyboard_id(void) {
     for (int i = 0; i < (int)(sizeof(strings) / sizeof(strings[0])); i++) {
         const char* str = strings[i];
         while (*str) {
-            hash ^= (uint32_t)(*str++);
+            hash ^= (uint32_t)(uint8_t)(*str++);
             hash *= FNV_PRIME;
         }
         hash ^= 0xFF;  // Separator between strings
